Simplify example tmux bindings, use if-shell
This makes use of an inlined variable in the `.tmux.conf` which removes
the duplication of the pattern, and `if-shell` in place of the inlined
command sequence previously used through `run-shell`.

Initially hoped to use via shell pattern matching, but that lost the
case-insensitivity that `grep -iqE` had.

Incorporates feedback from:

Suggest tmux if-shell instead of run-shell - #31
Less cluttered run-shell bindings - #44
Added suggestion for faster commands in zsh - #40