Add sane pane resizing to vim and tmux
Just a some config that allows tmux panes and vim splits to be resized universally with Meta/Alt + [h,j,k,l].  It's like vim-tmux-navigator on STEROIDS, so i thought this was the best place for it to call home.  As a new vimer & tmuxer, i was shocked that apparently nobody else has done this??? Maybe not the best idea for some reason? but it feels so good...