Prevent zooming out of tmux pane
There were issues (e.g. #56) and pull requests (#65 and #104) to improve this
behavior, but they were mainly rejected because it breaks other use cases.

This PR adds an option that can be enabled to have the current behavior
(`let g:tmux_navigator_move_out_of_zoomed_tmux = 1`) by default, but it also
recognizes when the user tries to get out of the zoomed pane twice, and
then allows for it.