Revert extra shell escaping in non-TPM bindings
Why
---

https://github.com/christoomey/vim-tmux-navigator/pull/181 was
introduced to fix an issues with the TPM install script not running
cleanly. This was due to the `C-\` key binding not being properly
escaped in the final line of the script.

In the process of fixing this in #181, it looks like we took it a bit
overboard and escaped _all_ `C-\` sequences. In the end, the others
should not be escaped, so this was causing issues for users not
configuring via TPM script.

This change
-----------

This change reverts all of the `C-\` escaping outside of the TPM script.