TmuxPaneCurrentCommand -> TmuxNavigatorProcessList
Previously we used the `{pane_current_command}` tmux variable and
matched against that, but a while back we switched to using the process
list to better handle some edge cases.

Overall that change has been a win, but the documentation has lagged a
bit around the debugging command `:TmuxPaneCurrentCommand` which is no
longer relevant.

This change replaces `:TmuxPaneCurrentCommand` with a new
`:TmuxNavigatorProcessList` which should allow us to debug the actual
issue related to the `ps` command.