Detect vim pane using more portable ps command.
The ps command on Cygwin or MSYS2 does not have the -o or -t flags so
fails to detect that vim is running in a pane. This version uses a
simplified ps command. It avoids the need for the -t flag by expanding
the tmux pane tty variable in the grep command to detect a running vim
instance.