Add DownOrRight and UpOrLeft directions.
From other tools I am used to switching windows/panes in a CCW way
where an 'up' translates to a 'left' if there is no window above and
a 'down' becomes a 'right' if there is no window below.

This patch adds two new commands

 TmuxNavigateDownOrRight
 TmuxNavigateUpOrLeft

that will check check for a neighbor in the requested direction and,
if there is none, change the direction.