Don't change active window when fzf is up
FZF uses Control+hjkl key combos for movement and it interferes with
this plugin. This commit adds the FZF process name(fzf) to the list of
programs that make this plugin not automatically trigger a window change.

This plugin also adds instructions to the readme file in case a user
wants to add their own programs to the list, or remove existing ones.

closes #346