TPM Plugin configure key mapping (#380)
as someone who loves TPM but uses a different key binding for pane navigation, I was a bit annoyed that the provided TPM plugin does not support configuring the key mapping for navigation. Of course, someone could resort to copying the provided setting to their .tmux.conf (or even fork this project), but I believe in using plugins to keep the config clean.

Thus, I've updated the plugin to support parameters for configuring the key mapping. It is possible to have multiple keys for each navigation action, as well as disable an individual mapping if desired. Have a look in the updated README to see how it works.

I've tested my changes in tmux version 3.2a, 3.4, as well as 2.9a (as the configuration parsing had some breaking changes in version 3.0). I hope I didn't miss any weird corner case.