TPM: Fix disabling key bindings
The documentation states that for the TPM plugin, key bindings could be
disabled by setting the corresponding option to an empty string.

This did not work. Instead `null` could be passed to disable a binding.

This commit changes it so that the function falls back to the default
value when the option is not set and allows for using an empty string to
disable a mapping. Also, `null` is kept for backward compatibility.