feat: faster vim detection
I noticed that the `ps` command is very slow (~200 ms) on one of my
machines. This makes pane switching slow both inside vim and in tmux.

This change uses `#{pane_current_command}` as a first attempt to
mitigate the slowdown, with a fallback to the current logic. This first
attempt handles most important cases (invoking `vim` directly or inside
an interactive shell).

Test: bash pattern-check
Test: manual: add `sleep 0.2` before `ps` to simulate the slowdown