refactor: unit tests return exit code
No change to logic. This refactors the unit tests to return a proper
exit code to indicate success or failure. This still runs all the unit
tests, but this will exit with status 0 if all tests have the expected
result (either "match" or "no match") and will exit with status 1 if at
least one test case has the unexpected result.

This script currently exits with status 1 because one test is already
broken (see issue #445).