refactor: remove redundant backslash sequence
I noticed that there is a `\\/` sequence in the regular expression. This
is totally redundant because `/` is not a special character in either
bash strings or in `grep -E` extended regular expressions. We can
simplify the regular expression slightly by changing this to `/`.

This does not remove any other double-backslash sequences. I learned in
issue #455 that these are necessary.

Test: bash pattern-check
Test: tmux kill-server and verify manually with multiple splits + nvim