test: introduce extra backslash evaluation in test (#454)
No change to production logic or behavior. I realized in #451 that there
is an extra round of backslash expansion which happens in the prod
environment but not the test environment. This updates the test
environment to also have an extra round evaluation to match. This change
should result in more reliable unit tests.

Issue #451
Test: bash pattern-check
Test: tmux kill-server and verify manually with multiple splits + nvim